+++ services/configService.ts
@@ -2,12 +2,14 @@ import { UserSettings } from '@/utils/types'
 import { storage } from '#imports'
 import { browser } from 'wxt/browser'
 import * as v from 'valibot'
+import { filterEmptyNamePairs } from '@/utils'
 
 export const defaultSettings: UserSettings = {
   names: {
     first: [],
     middle: [],
     last: [],
+    email: [],
   },
   enabled: true,
   blockContentBeforeDone: true,
@@ -16,6 +18,7 @@ export const defaultSettings: UserSettings = {
   highlightReplacedNames: true,
   syncSettingsAcrossDevices: false,
   theme: 'trans',
+  toggleKeybinding: null,
 }
 
 export async function getConfig(): Promise<UserSettings> {
@@ -35,8 +38,13 @@ export async function getConfig(): Promise<UserSettings> {
   return defaultSettings
 }
 
-export async function setConfig(config: UserSettings) {
-  const validatedConfig = v.parse(UserSettings, config)
+export async function setConfig(settings: UserSettings): Promise<void> {
+  const cleanedSettings = {
+    ...settings,
+    names: filterEmptyNamePairs(settings.names),
+  }
+
+  const validatedConfig = v.parse(UserSettings, cleanedSettings)
   const previousConfig = await getConfig()
 
   // Handle storage sync preference change